# server/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Laptop Store API",
    "version": "1.0.0",
    "description": "API documentation for Laptop Store"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/listing": {
      "post": {
        "tags": [
          "Listings"
        ],
        "summary": "Tạo một listing mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "category",
                  "roomCount",
                  "bathroomCount",
                  "guestCount",
                  "location",
                  "price",
                  "userId"
                ],
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Biệt thự sang trọng Đà Lạt"
                  },
                  "description": {
                    "type": "string",
                    "example": "View đồi thông, đầy đủ tiện nghi"
                  },
                  "imageSrc": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "uri"
                    },
                    "example": [
                      "https://example.com/image1.jpg"
                    ]
                  },
                  "category": {
                    "type": "string",
                    "example": "Villa"
                  },
                  "roomCount": {
                    "type": "integer",
                    "example": 3
                  },
                  "bathroomCount": {
                    "type": "integer",
                    "example": 2
                  },
                  "guestCount": {
                    "type": "integer",
                    "example": 6
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "value": {
                        "type": "string",
                        "example": "VN-63"
                      },
                      "label": {
                        "type": "string",
                        "example": "Lâm Đồng"
                      },
                      "flag": {
                        "type": "string",
                        "example": "🇻🇳"
                      },
                      "latlng": {
                        "type": "array",
                        "items": {
                          "type": "number"
                        },
                        "example": [
                          11.9404,
                          108.4583
                        ]
                      },
                      "region": {
                        "type": "string",
                        "example": "Tây Nguyên"
                      }
                    }
                  },
                  "price": {
                    "type": "number",
                    "example": 120
                  },
                  "userId": {
                    "type": "string",
                    "example": "64f2bcbcb20bd28a64a3c123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Listing đã được tạo thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "title": {
                      "type": "string"
                    },
                    "description": {
                      "type": "string"
                    },
                    "imageSrc": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "category": {
                      "type": "string"
                    },
                    "roomCount": {
                      "type": "integer"
                    },
                    "bathroomCount": {
                      "type": "integer"
                    },
                    "guestCount": {
                      "type": "integer"
                    },
                    "locationValue": {
                      "type": "string"
                    },
                    "userId": {
                      "type": "string"
                    },
                    "price": {
                      "type": "number"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string",
                          "format": "email"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy người dùng"
          }
        }
      },
      "get": {
        "tags": [
          "Listings"
        ],
        "summary": "Lấy danh sách tất cả listing theo điều kiện lọc",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "ID của người dùng (người tạo listing)"
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Danh mục của listing"
          },
          {
            "name": "roomCount",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Số phòng ngủ tối thiểu"
          },
          {
            "name": "bathroomCount",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Số phòng tắm tối thiểu"
          },
          {
            "name": "guestCount",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer"
            },
            "description": "Số lượng khách tối thiểu"
          },
          {
            "name": "localVale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Địa điểm cụ thể"
          },
          {
            "name": "price",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number"
            },
            "description": "Mức giá tối thiểu"
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Ngày bắt đầu dự kiến thuê"
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Ngày kết thúc dự kiến thuê"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách listing phù hợp",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "title": {
                        "type": "string"
                      },
                      "description": {
                        "type": "string"
                      },
                      "imageSrc": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "category": {
                        "type": "string"
                      },
                      "roomCount": {
                        "type": "integer"
                      },
                      "bathroomCount": {
                        "type": "integer"
                      },
                      "guestCount": {
                        "type": "integer"
                      },
                      "locationValue": {
                        "type": "string"
                      },
                      "userId": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      },
                      "user": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "email": {
                            "type": "string",
                            "format": "email"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reservation": {
      "post": {
        "tags": [
          "Reservations"
        ],
        "summary": "Tạo mới một reservation",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "listingId",
                  "startDate",
                  "endDate",
                  "totalPrice"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "ID của người dùng"
                  },
                  "listingId": {
                    "type": "string",
                    "description": "ID của listing"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "totalPrice": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Reservation được tạo thành công"
          },
          "400": {
            "description": "Dữ liệu đầu vào không hợp lệ"
          },
          "404": {
            "description": "Không tìm thấy user hoặc listing"
          }
        }
      }
    },
...
    "/api/user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Tạo người dùng mới",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "email",
                  "password"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Nguyễn Văn A"
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "example@gmail.com"
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "example": "abc123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "name": {
                      "type": "string",
                      "example": "Nguyễn Văn A"
                    },
                    "email": {
                      "type": "string",
                      "format": "email",
                      "example": "example@gmail.com"
                    },
                    "image": {
                      "type": "string",
                      "format": "uri",
                      "nullable": true,
                      "example": "https://example.com/avatar.jpg"
                    },
                    "favoriteIds": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "example": [
                        "abc123",
                        "def456"
                      ]
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email đã tồn tại"
          }
        }
      }
    },
...
    "/api/create-qr": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Tạo link thanh toán VNPay cho đơn đặt phòng",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "listingId",
                  "startDate",
                  "endDate",
                  "totalPrice"
                ],
                "properties": {
                  "listingId": {
                    "type": "string"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date"
                  },
                  "totalPrice": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Thành công, trả về URL thanh toán",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "paymentUrl": {
                      "type": "string",
                      "example": "https://sandbox.vnpayment.vn/paymentv2/vpcpay.html?..."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Không có token truy cập hợp lệ"
          },
          "500": {
            "description": "Lỗi server"
          }
        }
      }
    },
    "/api/vnpay_ipn": {
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "Xử lý callback từ VNPay sau thanh toán",
        "parameters": [
          {
            "in": "query",
            "name": "vnp_TxnRef",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vnp_Amount",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vnp_ResponseCode",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vnp_TransactionNo",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vnp_BankCode",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vnp_CardType",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vnp_PayDate",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "vnp_SecureHash",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kết quả xử lý IPN VNPay",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "RspCode": {
                      "type": "string"
                    },
                    "Message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}